+++ org.emoflon.roam.core/src/org/emoflon/roam/core/RoamTypeConstraint.java
@@ -27,8 +27,8 @@ public abstract class RoamTypeConstraint <CONTEXT extends EObject> extends RoamC
 	
 	@Override
 	public ILPConstraint<Integer> buildConstraint(final CONTEXT context) {
-		Integer constTerm = buildConstantTerm(context);
-		List<ILPTerm<Integer, Integer>> terms = buildVariableTerms(context);
+		double constTerm = buildConstantTerm(context);
+		List<ILPTerm<Integer, Double>> terms = buildVariableTerms(context);
 		return new ILPConstraint<Integer>(constTerm, constraint.getExpression().getOperator(), terms);
 	}
